requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.36
psycopg[binary]
pydantic[email]==2.9.2
alembic==1.13.1
python-dotenv==1.0.1
pydantic-settings
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.1.1
fastapi-mail==1.4.1
python-multipart==0.0.9
fastapi-limiter==0.1.5
redis==4.6.0
cloudinary==1.40.0
pytest==8.3.2
pytest-cov==5.0.0
pytest-asyncio==0.23.8
sphinx==7.3.7
sphinx-autobuild==2024.4.16
fakeredis==2.23.2
pydantic==2.9.2
bcrypt==4.0.1
httpx==0.27.2